backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
fastapi==0.135.1
gitdb==4.0.12
GitPython==3.1.46
groq==1.1.1
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
numpy==2.4.3
portalocker==3.2.0
protobuf==7.34.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
qdrant-client==1.17.1
requests==2.32.5
smmap==5.0.3
sniffio==1.3.1
starlette==0.52.1
tree-sitter-languages==1.10.2
tree_sitter==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
